+++ src/test/java/org/cojen/maker/SwitcherTest.java
@@ -17,7 +17,9 @@
 package org.cojen.maker;
 
 import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
 
+import java.time.DayOfWeek;
 import java.time.LocalTime;
 import java.time.Month;
 
@@ -65,6 +67,27 @@ public class SwitcherTest {
             fail();
         } catch (NullPointerException e) {
             assertTrue(e.getMessage().contains("Case cannot be null"));
+        }
+
+        try {
+            mm.param(0).switch_(def, new Object[] {mm.param(0)}, mm.label());
+            fail();
+        } catch (IllegalArgumentException e) {
+            assertTrue(e.getMessage().contains("Case isn't a constant"));
+        }
+
+        try {
+            MethodMaker mm2 = cm.addMethod(int.class, "t2", String.class).public_().static_();
+            mm2.param(0).switch_(mm2.label(), new Month[] {Month.JUNE}, mm.label());
+        } catch (IllegalStateException e) {
+            assertTrue(e.getMessage().contains("Not switching on an Enum"));
+        }
+
+        try {
+            MethodMaker mm2 = cm.addMethod(int.class, "t2", DayOfWeek.class).public_().static_();
+            mm2.param(0).switch_(mm2.label(), new Month[] {Month.JUNE}, mm.label());
+        } catch (IllegalArgumentException e) {
+            assertTrue(e.getMessage().contains("Enum case"));
         }
     }
 
@@ -124,13 +147,11 @@ public class SwitcherTest {
 
         assertEquals(-1, m.invoke(null, "xxxxxxxxx"));
 
-        if (keys.length > 0) {
-            try {
-                m.invoke(null, (String) null);
-                fail();
-            } catch (InvocationTargetException e) {
-                assertTrue(e.getCause() instanceof NullPointerException);
-            }
+        try {
+            m.invoke(null, (String) null);
+            fail();
+        } catch (InvocationTargetException e) {
+            assertTrue(e.getCause() instanceof NullPointerException);
         }
     }
 
@@ -222,5 +243,173 @@ public class SwitcherTest {
         }
 
         assertEquals(-1, m.invoke(null, 999L));
+    }
+
+    @Test
+    public void enums() throws Exception {
+        doEnums();
+    }
+
+    @Test
+    public void enumsNoSwitchBootstraps() throws Exception {
+        Switcher.NO_SWITCH_BOOTSTRAPS = true;
+        try {
+            doEnums();
+        } finally {
+            Switcher.NO_SWITCH_BOOTSTRAPS = false;
+        }
+    }
+
+    private void doEnums() throws Exception {
+        doEnums(new Month[0]);
+        doEnums(Month.FEBRUARY, Month.MARCH);
+        doEnums(Month.FEBRUARY, Month.MARCH, Month.JULY);
+    }
+
+    private void doEnums(Month... keys) throws Exception {
+        doEnums(ClassMaker.begin(), keys);
+    }
+
+    @Test
+    public void enumsExternal() throws Exception {
+        doEnumsExternal("_Enums$$$");
+    }
+
+    @Test
+    public void enumsExternalNoSwitchBootstraps() throws Exception {
+        Switcher.NO_SWITCH_BOOTSTRAPS = true;
+        try {
+            doEnumsExternal("_EnumsNoBoot$$$");
+        } finally {
+            Switcher.NO_SWITCH_BOOTSTRAPS = false;
+        }
+    }
+
+    private void doEnumsExternal(String suffix) throws Exception {
+        String base = SwitcherTest.class.getName() + suffix;
+
+        doEnums(ClassMaker.beginExternal(base + 1), new Month[0]);
+        doEnums(ClassMaker.beginExternal(base + 2), Month.FEBRUARY, Month.MARCH);
+        doEnums(ClassMaker.beginExternal(base + 3), Month.FEBRUARY, Month.MARCH, Month.JULY);
+    }
+
+    private void doEnums(ClassMaker cm, Month... keys) throws Exception {
+        cm.public_();
+
+        MethodMaker mm = cm.addMethod(int.class, "map", Month.class).public_().static_();
+
+        Label[] labels = new Label[keys.length];
+        for (int i=0; i<labels.length; i++) {
+            labels[i] = mm.label();
+        }
+
+        Label notFound = mm.label();
+
+        mm.param(0).switch_(notFound, keys, labels);
+
+        for (int i=0; i<labels.length; i++) {
+            labels[i].here();
+            mm.return_(i + 100);
+        }
+
+        notFound.here();
+        mm.return_(-1);
+
+        var m = cm.finish().getMethod("map", Month.class);
+
+        for (int i=0; i<keys.length; i++) {
+            assertEquals(i + 100, m.invoke(null, keys[i]));
+        }
+
+        assertEquals(-1, m.invoke(null, Month.JANUARY));
+
+        try {
+            m.invoke(null, (Month) null);
+            fail();
+        } catch (InvocationTargetException e) {
+            assertTrue(e.getCause() instanceof NullPointerException);
+        }
+    }
+
+    @Test
+    public void enumsExternalNoSwitchBootstrapsRecycle() throws Exception {
+        // Tests that only one inner class is generated per enum type.
+
+        Class<?> clazz;
+
+        Switcher.NO_SWITCH_BOOTSTRAPS = true;
+        try {
+            ClassMaker cm = ClassMaker.beginExternal(SwitcherTest.class.getName() + "_Recycle$$$");
+            cm.public_();
+
+            {
+                MethodMaker mm = cm.addMethod(int.class, "m1", Month.class).public_().static_();
+                Month[] keys = {Month.FEBRUARY, Month.MAY};
+                Label[] labels = {mm.label(), mm.label()};
+                Label notFound = mm.label();
+                mm.param(0).switch_(notFound, keys, labels);
+                for (int i=0; i<labels.length; i++) {
+                    labels[i].here();
+                    mm.return_(i + 100);
+                }
+                notFound.here();
+                mm.return_(-1);
+            }
+
+            {
+                MethodMaker mm = cm.addMethod(int.class, "m2", Month.class).public_().static_();
+                Month[] keys = {Month.JUNE, Month.FEBRUARY};
+                Label[] labels = {mm.label(), mm.label()};
+                Label notFound = mm.label();
+                mm.param(0).switch_(notFound, keys, labels);
+                for (int i=0; i<labels.length; i++) {
+                    labels[i].here();
+                    mm.return_(i + 200);
+                }
+                notFound.here();
+                mm.return_(-1);
+            }
+
+            {
+                MethodMaker mm = cm.addMethod(int.class, "m3", DayOfWeek.class).public_().static_();
+                DayOfWeek[] keys = {DayOfWeek.FRIDAY, DayOfWeek.MONDAY};
+                Label[] labels = {mm.label(), mm.label()};
+                Label notFound = mm.label();
+                mm.param(0).switch_(notFound, keys, labels);
+                for (int i=0; i<labels.length; i++) {
+                    labels[i].here();
+                    mm.return_(i + 300);
+                }
+                notFound.here();
+                mm.return_(-1);
+            }
+
+            clazz = cm.finish();
+        } finally {
+            Switcher.NO_SWITCH_BOOTSTRAPS = false;
+        }
+
+        assertEquals(2, clazz.getDeclaredClasses().length);
+
+        {
+            Method m = clazz.getMethod("m1", Month.class);
+            assertEquals(100, m.invoke(null, Month.FEBRUARY));
+            assertEquals(101, m.invoke(null, Month.MAY));
+            assertEquals(-1, m.invoke(null, Month.DECEMBER));
+        }
+
+        {
+            Method m = clazz.getMethod("m2", Month.class);
+            assertEquals(200, m.invoke(null, Month.JUNE));
+            assertEquals(201, m.invoke(null, Month.FEBRUARY));
+            assertEquals(-1, m.invoke(null, Month.DECEMBER));
+        }
+
+        {
+            Method m = clazz.getMethod("m3", DayOfWeek.class);
+            assertEquals(300, m.invoke(null, DayOfWeek.FRIDAY));
+            assertEquals(301, m.invoke(null, DayOfWeek.MONDAY));
+            assertEquals(-1, m.invoke(null, DayOfWeek.SUNDAY));
+        }
     }
 }