+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 v2.5.4
 ------
 * Added a method to insert code after a label.
+* Added a switch method which supports enum cases.
 * Added a switch method which supports object cases.
 
 v2.5.3 (2023-12-12)